Some folder and code cleaning and restructure
