--- src/Components/Footer/Footer.css
@@ -1,22 +1,0 @@
-.card-footer {
-    /*margin-top: 1rem;*/
-    /*margin-bottom: 1rem;*/
-    font-size: 0.8rem;
-    background-color: transparent;
-    border: 0;
-
-
-
-
-    position: absolute;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    text-align: center;
-}
-
-@media only screen and (max-width: 411px) {
-    .card-footer span {
-        font-size: .7rem;
-    }
-}
