+++ src/Parts/Header.css
@@ -1,5 +1,4 @@
 .Header.navbar {
-    /*background-color: #000000;*/
     background-color: #1d1d1d;
 }
